Expand try_perform! not to call perform!
The reason for that is the RSpec action matcher stubbing making an
expectation that when `try_perform!` is expected to be called,
`perform!` won't be called. The stubbing works fine except for the case
when `and_call_original` modifier is added, and the `perform!` is
actually called.

There is an additional edge case when we expect that `perform!` would be
called, and it is called, but indirectly, via `try_perform!` that is
undesirable.